# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "skip_verification": {
            "default": true,
            "title": "Skip Verification",
            "type": "boolean"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "use_llm": {
            "default": true,
            "title": "Use Llm",
            "type": "boolean"
          },
          "use_rag": {
            "default": false,
            "title": "Use Rag",
            "type": "boolean"
          },
          "whisper_model": {
            "default": "medium",
            "pattern": "^(tiny|base|small|medium|large)$",
            "title": "Whisper Model",
            "type": "string"
          }
        },
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "**한국어 사기 신호 검출 reference implementation** — 음성·텍스트·이미지·PDF·URL 입력에서 위험 신호를 검출하고 각 신호의 학술/법적 근거를 transparent 하게 보고한다.\n\n⚠️ **Identity Boundary** (CLAUDE.md): ScamGuardian 은 *사기 판정을 내리지 않는다*. VirusTotal 이 70개 백신의 검출 결과를 보고만 하는 모델과 동일. 판정 logic 은 통합한 기업(통신사·은행·메신저 앱)이 자기 risk tolerance 에 따라 구현한다.\n\n- **Public 4 endpoint** (`Public` 태그) — 외부 통합 시작점, `DetectionReport` 응답\n- **Webhook** (`Webhook` 태그) — 카카오 오픈빌더\n- **Admin** (3 태그) — 라벨링 / platform / training\n- **v4 draft** — Live Call Guard 설계 preview (구현 X)\n\n통합 가이드: [`docs/INTEGRATION_GUIDE.md`](https://github.com/example/scamguardian-v2/blob/main/docs/INTEGRATION_GUIDE.md)",
    "title": "ScamGuardian Signal Detection API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analyze": {
      "post": {
        "description": "ScamGuardian 의 메인 검출 엔드포인트. 한국어 텍스트 또는 URL/YouTube/음성·영상·이미지·PDF 파일 경로를 받아 Phase 0 (안전성) → Phase 0.5 (sandbox) → Phase 1 (STT/OCR) → Phase 2-5 (분류·추출·검증·검출) 전체 파이프라인을 수행한다.\n\n⚠️ **Identity Boundary** (CLAUDE.md): ScamGuardian 은 **사기 판정을 내리지 않는다**. 검출된 위험 신호 list 와 각 신호의 학술/법적 근거만 transparent 하게 보고한다. VirusTotal 이 70개 백신의 검출 결과를 보고만 하는 모델과 동일. 최종 판정 logic 은 통합 기업이 자체 risk tolerance 에 따라 구현한다.\n\n**Request body** (`AnalyzeRequest`):\n- `source` *or* `text` — 둘 중 하나는 필수. URL 도 `source` 로 전달\n- `whisper_model` — `tiny|base|small|medium|large` (음성 입력 시에만 영향)\n- `skip_verification` — true 면 Phase 4 (Serper) 건너뜀, 응답 빨라짐\n- `use_llm` — 무시됨 (서버에서 강제 true)\n- `use_rag` — 과거 라벨 사례 RAG 검색 여부\n\n**응답 핵심 필드** (`DetectionReport`):\n- `scam_type` — 분류된 스캠 유형 (검출 컨텍스트, 판정 X)\n- `detected_signals` — `[{flag, label_ko, rationale, source, evidence, description, detection_source}]` 검출된 위험 신호 list. 각 신호마다 학술/법적 근거(`rationale`) 와 출처(`source`) 포함\n- `summary` — `\"위험 신호 N개 검출되었습니다. 자세한 근거는 detected_signals 참고.\"`\n- `disclaimer` — 통합 기업 판정 logic 안내 문구\n- `entities` — `[{label, text, score, source}]` 추출된 엔티티\n- `transcript_text` — 본문 (URL/파일 입력 시 STT/OCR 결과)\n- `analysis_run_id` — DB 저장된 경우 UUID\n\n❌ 응답에 `total_score`, `risk_level`, `is_scam`, `agent_verdict` 등 *판정* 필드는 **없다**.\n\n**인증**: API key 필수 (`Authorization: Bearer sg_...` 또는 `X-API-Key`).\n\n**에러**:\n- `400` — 빈 입력 / 어뷰즈 가드 reject (REPETITIVE / GIBBERISH / DUPLICATE)\n- `401` — API key 누락 또는 무효\n- `403` — API key revoked\n- `423` — 어뷰즈 가드 BLOCKED (위반 누적 차단)\n- `429` — RPM 또는 월 USD/호출 cap 초과 (`Retry-After` 헤더)\n- `500` — 파이프라인 내부 오류\n\n**curl** (텍스트):\n```bash\ncurl -X POST https://api.example.com/api/analyze \\\n  -H \"Authorization: Bearer sg_xxx\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\"text\": \"검찰청입니다. 즉시 300만원 송금하세요.\"}'\n```\n\n**curl** (YouTube URL):\n```bash\ncurl -X POST https://api.example.com/api/analyze \\\n  -H \"Authorization: Bearer sg_xxx\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\"source\": \"https://youtu.be/abcd1234\"}'\n```",
        "operationId": "analyze_api_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Analyze Api Analyze Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "빈 입력 / 어뷰즈 reject / 환경 미설정"
          },
          "401": {
            "description": "API key 누락 또는 무효"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "423": {
            "description": "어뷰즈 누적 차단 — 1시간 후 재시도"
          },
          "429": {
            "description": "Rate limit 초과 — `Retry-After` 헤더 참조"
          }
        },
        "summary": "텍스트·URL 분석 — 위험 신호 검출 (Signal Detection)",
        "tags": [
          "Public"
        ]
      }
    },
...
    "/api/result/{token}": {
      "get": {
        "description": "카카오 챗봇 응답 카드의 '자세히 보기' 링크가 호출하는 엔드포인트. 결과 발급 시점에 1시간 유효 토큰을 발급(`secrets.token_urlsafe(16)`)하고 메모리에 저장한다. 인메모리 저장이라 서버 재시작 시 토큰은 모두 만료.\n\n**응답**:\n- `result` — `DetectionReport.to_dict()` 전체 (`/api/analyze` 와 동일 schema)\n- `user_context` — 사용자가 챗봇과 나눈 Q&A 요약 (있을 때)\n- `input_type` — `TEXT|URL|VIDEO|FILE|IMAGE|PDF`\n- `chat_history` — 봇/사용자 대화 시간순 전체\n- `flag_rationale` — 검출된 신호별 `{rationale, source}` 매핑\n- `expires_at` — 토큰 만료 시각 (epoch sec)\n\n**인증**: 선택. 토큰 자체가 인증 역할 (URL 알면 누구나 조회).\n\n**에러**:\n- `404` — 토큰이 존재한 적 없음 또는 서버 재시작\n- `410` — 토큰 만료 (1시간 경과)\n\n**curl**:\n```bash\ncurl https://api.example.com/api/result/abcd1234efgh5678\n```",
        "operationId": "get_result_by_token_api_result__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Result By Token Api Result  Token  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "토큰 없음"
          },
          "410": {
            "description": "토큰 만료 — 결과 재요청 필요"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "결과 토큰으로 분석 결과 조회",
        "tags": [
          "Public"
        ]
      }
    },
...
  },
...
}